requirements.txt
numpy
pandas
scikit-learn
matplotlib
python-dotenv
mysql-connector-python
pymysql

# -e .